First attempt to simplify export of fake libc symbols (wrong way).
Now libc-ns3.h defines all the functions that should be exported, along
with the way the function should be exported: either it is just
redirected to a native call NATIVE(func) or redirected to an appropriate
DCE call: DCE(func).  By convention, DCE redicted calls are actually
redirected to the function with the same prototype as libc function, but
name of the dce function starts with `dce_' prefix.