Implementing a new gcc-specific way to redirect calls to dce/native libc function.
The method uses gcc extension for constructing function calls
(http://tigcc.ticalc.org/doc/gnuexts.html#SEC67):
- __builtin_apply_args
- __builtin_apply
_ __builtin_return

The only problem right now is that __builtin_apply needs a parameter
that represents the size of the stack argument data, which according to
docs is not easy to compute. However, with arbitrary large number (256)
everything seems to work... until a dce app uses printf-like calls with
thousands of parameters, which is highly unlikely.